+++ modules/clientutils.js
@@ -407,11 +407,20 @@
                 this.log("mouseEvent(): Couldn't find any element matching '" + selector + "' selector", "error");
                 return false;
             }
-            var evt = document.createEvent("MouseEvents");
-            evt.initMouseEvent(type, true, true, window, 1, 1, 1, 1, 1, false, false, false, false, 0, elem);
-            // dispatchEvent return value is false if at least one of the event
-            // handlers which handled this event called preventDefault
-            return elem.dispatchEvent(evt);
+            try {
+                var evt = document.createEvent("MouseEvents");
+                evt.initMouseEvent(type, true, true, window, 1, 1, 1, 1, 1, false, false, false, false, 0, elem);
+                // dispatchEvent return value is false if at least one of the event
+                // handlers which handled this event called preventDefault;
+                // so we cannot returns this results as it cannot accurately informs on the status
+                // of the operation
+                // let's assume the event has been sent ok it didn't raise any error
+                elem.dispatchEvent(evt);
+                return true;
+            } catch (e) {
+                this.log("Failed dispatching " + type + "mouse event on " + selector + ": " + e, "error");
+                return false;
+            }
         };
 
         /**